Final_Project/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
distlib==0.3.6
filelock==3.9.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-client==0.0.1
mysqlclient==2.2.0
platformdirs==3.0.0
pyFirmata==1.1.0
PyJWT==2.8.0
pyserial==3.5
SQLAlchemy==2.0.20
typing_extensions==4.7.1
virtualenv==20.19.0
Werkzeug==2.3.7
